Refactor and validate the scalar production from hadron decays.
* Validate production of the scalar through 2-body exclusive hadronic decays.
* Add unit-tests to compare the numbers to the Mathematica implementation.

NOTE: There is an ambiguity regarding which renormalization scheme / scale to
use for the calculation of the ξ factors. This leads to enourmous uncertainties
(~10) on the final decay rate due to the top loop.
Indeed, Γ ∝ mt^4, with Mt = 173 GeV (on-shell) vs. m̄t(μ=2 GeV) ~ 300 GeV (MS).
Here, I have chosen to use the on-shell mass, which results in a more
conservative sensitivity estimate.