Use the correct heavy quark masses to compute the ξ factors.
Switch from using the on-shell mass to the scale-invariant MS-bar mass for the
heavy quarks, since they must be manually decoupled in order to avoid large
logarithms and unphysical corrections to the beta function in the low-energy
perturbation theory.

Thanks to Prof. Aneesh Manohar to help us understand this calculation.

Unit-tests for exclusive two-body hadronic production have been updated.