Add 'S -> K K' channels and `TwoKaons` class.
Channels added:
* 'S -> K0 Kbar0'
* 'S -> K+ K-'

The total decay width is evaluated using tabulated data from Winkler [1], up to
a Scalar mass of 2 GeV, above which this calculation is expected to break down.
The 1.4–2 GeV mass range relies on an extrapolation is is less accurate, due to
additional hadronic channels opening (4π, ηη, ρρ, …).

The tabulated data is linearly interpolated, since higher order schemes lead to
severe ringing with this data set. A better alternative might be to perform the
interpolation in log-space.

The `TwoKaons` class only implements `normalized_width`, and relies on the
default implementation for other methods.

[1] Winkler, M.W., 2019.
    Decay and Detection of a Light Scalar Boson Mixing with the Higgs.
    Physical Review D 99. https://doi.org/10.1103/PhysRevD.99.015018
    Data extracted from Fig. 4.