Allow specifying multiple coupling coefficients.
**API BREAKING**

Instead of specifying a unique coupling θ as a number or array, the relevant
methods now accept a dictionary of named couplings (the values of which are
themselves numbers or arrays).

Unit-tests and examples have been adapted to use the new interface.

To switch to the new interface, make the following replacement:
    `obj.method(mS, theta)` -> `obj.method(mS, {'theta': theta, 'alpha': 0})`
Some high-level methods now additionally provide the following interface:
    `obj.method(mS, theta=theta, alpha=0)`