Use a different plot for the production widths and BRs for the quartic coupling.
The scales and couplings are different for the mixing with the SM Higgs and the
quartic coupling, so it makes sense to use two different plots.

This also makes each plot more readable, and the lower resolution used for the
quartic coupling allows to considerably reduce the run time for the test suite.